requirements.txt
selenium
chromedriver_autoinstaller
pyyaml
wget